replace setInterval with node-cron
