+++ devmgmtV2/package.json
@@ -15,6 +15,7 @@
     "diskspace": "^2.0.0",
     "express": "^4.16.2",
     "image-size": "^0.6.2",
+    "node-cron": "^1.2.1",
     "os-utils": "0.0.14",
     "q": "^1.5.1",
     "unzip": "^0.1.11"
